@@ -1,11 +1,20 @@
-module github.com/IBM-Cloud/hpcs-grep11-go
+module github.com/IBM-Cloud/hpcs-grep11-go/v2
 
-go 1.15
+go 1.21
 
 require (
+	github.com/Zilliqa/gozilliqa-sdk/v3 v3.0.0-20230120123140-369a05d8c3b8
+	github.com/btcsuite/btcd/btcec/v2 v2.2.0
 	github.com/gogo/protobuf v1.3.2
-	github.com/golang/protobuf v1.4.3
-	golang.org/x/net v0.0.0-20201021035429-f5854403a974 // indirect
-	google.golang.org/genproto v0.0.0-20201001141541-efaab9d3c4f7 // indirect
+	github.com/golang/protobuf v1.5.2
 	google.golang.org/grpc v1.34.0
+)
+
+require (
+	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.0.1 // indirect
+	golang.org/x/net v0.0.0-20201021035429-f5854403a974 // indirect
+	golang.org/x/sys v0.3.0 // indirect
+	golang.org/x/text v0.3.3 // indirect
+	google.golang.org/genproto v0.0.0-20201001141541-efaab9d3c4f7 // indirect
+	google.golang.org/protobuf v1.26.0 // indirect
 )
